Moved the condense functions
