+++ CalcExpr/Parsing/Enums.cs
@@ -1,0 +1,10 @@
+﻿namespace CalcExpr.Parsing;
+
+public enum Brackets
+{
+    None = 0,
+    Parenthesis = 1,
+    Square = 2,
+    Curly = 4,
+    All = 0xFF,
+}
